cheri: io_uing: Add casts, allow with CHERI_KERNEL
Add neccessary casts between uintptr_t and unsigned long to io_uring/
and allow io_uring to be configured in a CHERI_KERNEL again.

+++ include/linux/io_uring_types.h
@@ -58,7 +58,7 @@ struct io_wq_work {
 
 struct io_fixed_file {
 	/* file * with additional FFS_* flags */
-	unsigned long file_ptr;
+	uintptr_t file_ptr;
 };
 
 struct io_file_table {
